bringing the opeartions back to bytes
